Moved input reading funcitons to Schedule and made allClasses static in schedule for more effeciency. Also made all properties read only. This is to prevent modification of variables by anything other than functions in schedule.
